Bug fix when deleting account

+++ app/Http/Controllers/AdminController.php
@@ -84,11 +84,25 @@ class AdminController extends Controller
      */
     public function destroy(User $user)
     {
-        if (file_exists(public_path() .'/images/'. $user->photo->name)) {//kontrollo nese ekziston foto ne storage para se te fshihet
-            unlink(public_path() .'/images/'. $user->photo->name);
+        if ($user->photo_id != 1){
+            if (file_exists(public_path() .'/images/'. $user->photo->name)) {//kontrollo nese ekziston foto ne storage para se te fshihet
+                unlink(public_path() .'/images/'. $user->photo->name);
+            }
         }
+        $user->name = $user->name . " " . $user->surname;
+        $user->surname = null;
+        $user->gender = null;
+        $user->username = null;
+        $user->slug = null;
+        $user->university_id = null;
+        $user->city_id = null;
+        $user->about = null;
+        $user->email = null;
+        $user->password = null;
+        $user->photo_id = 1;
+        $user->role_id = 1;
 
-        $user->delete();
+        $user->save();
 
 
         return back();
